[UPD] refactor view module
