+++ src/Core/Controller.php
@@ -1,9 +1,12 @@
 <?php
 /*
- * Copyright (c) 2023. wepesi dev framework
+ * Copyright (c) 2023-2024. Wepesi Dev Framework
  */
 
 namespace Wepesi\Core;
+
+use Wepesi\Core\Views\Provider\Contract\ViewsContract;
+use Wepesi\Core\Views\View;
 
 /**
  *
@@ -13,7 +16,7 @@ abstract class Controller
     /**
      * @var View
      */
-    protected View $view;
+    protected ViewsContract $view;
 
     /**
      *
