+++ src/Core/Views/View.php
@@ -1,18 +1,25 @@
 <?php
+/*
+ * Copyright (c) 2024. Wepesi Dev Framework
+ */
 
-namespace Wepesi\Core;
+namespace Wepesi\Core\Views;
 
-use \DOMDocument;
-use \DOMXPath;
+use DOMDocument;
+use DOMXPath;
 use Exception;
+use Wepesi\Core\Application;
+use Wepesi\Core\Escape;
 use Wepesi\Core\Http\Response;
+use Wepesi\Core\MetaData;
+use Wepesi\Core\Views\Provider\ViewBuilderProvider;
 use function libxml_clear_errors;
 use function libxml_use_internal_errors;
 
 /**
  *
  */
-class View
+class View extends ViewBuilderProvider
 {
     /**
      *
@@ -21,23 +28,16 @@ class View
     /**
      * @var array
      */
-    private static array $jslink = [];
+    private static array $js_link_path = [];
     /**
      * @var array
      */
-    private static array $stylelink = [];
+    private static array $style_link_path = [];
     /**
      * @var string|null
      */
     private static ?string $metadata = null;
-    /**
-     * @var array
-     */
-    private array $data = [];
-    /**
-     * @var string
-     */
-    private string $folder_name = '';
+
     /**
      * @var string
      */
@@ -48,27 +48,27 @@ class View
     private string $layout_content = '';
 
     /**
-     * @param string $js_link
-     *
+     * Provide js link to be set up on the page header
+     * @param string $path file path
+     * @param bool $external set to true for an external link
      * @return void
      */
-    public static function setJsToHead(string $js_link, bool $external = false)
+    public static function setJsToHead(string $path, bool $external = false)
     {
-        self::$jslink[] = [
-            'link' => $js_link,
-            'external' => $external
-        ];
+        self::$js_link_path[] = $path;
     }
 
     /**
-     * @param string $style_link
+     * Provide css link path to set up stylesheet on the page header
+     *
+     * @param string $path
      *
      * @return void
      * add
      */
-    public static function setStyleToHead(string $style_link)
+    public static function setStyleToHead(string $path)
     {
-        self::$stylelink[] = $style_link;
+        self::$style_link_path[] = $path;
     }
 
     /**
@@ -79,15 +79,6 @@ class View
     public static function setMetaData(MetaData $metadata)
     {
         self::$metadata = $metadata->build();
-    }
-
-    /**
-     * @param string $folder_name
-     * @return void
-     */
-    public function setFolder(string $folder_name)
-    {
-        $this->folder_name = Escape::addSlashes($folder_name);
     }
 
     /**
@@ -122,6 +113,7 @@ class View
     }
 
     /**
+     * Render content from prided file
      * @param string $view
      *
      * @return false|string|void
@@ -209,15 +201,14 @@ class View
             $head = $xpath->query('//head/title');
             $template = $dom->createDocumentFragment();
             // add a style link to the head tag of the page
-            foreach (self::$stylelink as $k => $v) {
+            foreach (self::$style_link_path as $k => $v) {
                 $template->appendXML('<link rel="stylesheet" type="text/css" href="' . $v . '">');
                 $head[0]->parentNode->insertbefore($template, $head[0]->nextSibling);
             }
             // add a script link to the head of the page
-            foreach (self::$jslink as $k => $v) {
+            foreach (self::$js_link_path as $k => $v) {
                 $link = $v['link'];
                 $src = '<script src="' . $link . '" type="text/javascript"></script>';
-                if (!$v['external']) $src = Bundles::insertJS($link, false, true);
                 $template->appendXML($src);
                 $head[0]->parentNode->insertbefore($template, $head[0]->nextSibling);
             }
@@ -234,25 +225,9 @@ class View
         }
     }
 
-    /**
-     * assign variables data to be displayed on file_page
-     *
-     * @param string $variable
-     * @param        $value
-     */
-    public function assign(string $variable, $value)
-    {
-        $this->data[$variable] = $value;
-    }
 
-    /**
-     * List all data assigned before being displayed
-     * @return array
-     */
-    public function getAssignData(): array
-    {
-        return $this->data;
-    }
+
+
     /**
      * you should provide the extension of your file,
      * in another case the file will be missing
